+++ module/test-dialog.js
@@ -163,6 +163,9 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
     const myActor = fromUuidSync(this.test.actor)
     context.test.hasModifiers = false;
 
+    if (typeof this.test.DNDescriptor === 'number') {
+      context.numberDN = { value: this.test.DNDescriptor }
+    }
     // The wound penalties are never more than -3, regardless on how many wounds a token can suffer / have. CrB p. 117
     context.test.woundModifier = -Math.min(myActor.system.wounds.value ?? 0, 3);